+++ OrbitGl/SamplingReport.h
@@ -19,7 +19,7 @@ public:
     std::shared_ptr< class SamplingProfiler > GetProfiler() const { return m_Profiler; }
     const std::vector< std::shared_ptr<class DataView> > & GetThreadReports() { return m_ThreadReports; }
     void SetCallstackDataView( class CallStackDataView* a_DataView ){ m_CallstackDataView = a_DataView; }
-    void OnSelectAddress( unsigned long long a_Address, unsigned long a_ThreadId );
+    void OnSelectAddress( uint64_t a_Address, uint32_t a_ThreadId );
     void OnCallstackIndexChanged( int a_Index );
     void IncrementCallstackIndex();
     void DecrementCallstackIndex();
